added repositories and fixed some errors and imports

+++ back-end/src/main/java/ma/inpt/cedoc/repositories/utilisateursRepositories/UtilisateurRepository.java
@@ -1,7 +1,8 @@
 package ma.inpt.cedoc.repositories.utilisateursRepositories;
 
-import ma.inpt.cedoc.model.entities.utilisateurs.Utilisateur;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UtilisateurRepository extends JpaRepository<Utilisateur, Integer> {
+import ma.inpt.cedoc.model.entities.utilisateurs.Utilisateur;
+
+public interface UtilisateurRepository extends JpaRepository<Utilisateur, Long> {
 }
